[CI] Fix triggers of docs builds and condition for setting `BREEZE_LABEL_BUILD_ALL_EXAMPLES` (#346)
We aren't interested in the `unlabeled` (removing any labele shouldn't trigger a
new build) and `edited` (editing title or body of the PR shouldn't trigger a new
build) events at all.  Also, when the PR is `labeled`, we are only interested in
when we are adding the special label, not anything else.

[skip ci]